requirements.txt
# ============================================================================
# Medical RAG Knowledge Base - Dependencies
# Python 3.11+ | Optimized for RTX 3050 4GB VRAM
# Package Manager: uv (fast, reliable)
# ============================================================================

# ---------------------------------------------------------------------------
# CORE VECTOR DATABASE & EMBEDDINGS
# ---------------------------------------------------------------------------
chromadb==0.4.22
sentence-transformers==3.3.1
huggingface-hub==0.26.5

# ---------------------------------------------------------------------------
# NLP & TEXT PROCESSING
# ---------------------------------------------------------------------------
spacy==3.7.4
# Note: After install, run: python -m spacy download en_core_web_sm

# BM25 for hybrid search
rank-bm25==0.2.2

# ---------------------------------------------------------------------------
# DEEP LEARNING & NUMERICAL COMPUTING
# ---------------------------------------------------------------------------
torch>=2.0.0
numpy==1.26.3
scikit-learn==1.4.0

# ---------------------------------------------------------------------------
# CONFIGURATION & SERIALIZATION
# ---------------------------------------------------------------------------
toml==0.10.2

# ---------------------------------------------------------------------------
# UTILITIES
# ---------------------------------------------------------------------------
tqdm==4.66.1
requests==2.31.0

# Rich terminal formatting for query interface
rich==13.7.0

# ============================================================================
# INSTALLATION INSTRUCTIONS (with uv)
# ============================================================================
#
# 1. Activate your virtual environment:
#    .\RAG\Scripts\activate  (Windows PowerShell)
#
# 2. Install with uv (recommended):
#    uv pip install -r requirements.txt
#
#    OR with standard pip:
#    pip install -r requirements.txt
#
# 3. Download spaCy language model:
#    python -m spacy download en_core_web_sm
#
# 4. Verify GPU detection (optional):
#    python -c "import torch; print(f'CUDA Available: {torch.cuda.is_available()}')"
#    python -c "import torch; print(f'GPU: {torch.cuda.get_device_name(0) if torch.cuda.is_available() else \"N/A\"}')"
#
# ============================================================================
# OPTIONAL: LLM INTEGRATION (uncomment if needed)
# ============================================================================
#
# # LangChain for RAG orchestration
# langchain==0.1.6
# langchain-community==0.0.20
#
# # OpenAI API client
# openai==1.10.0
#
# # Ollama for local LLMs
# ollama==0.1.6
#
# # HuggingFace Transformers
# transformers==4.37.2
# accelerate==0.26.1
# bitsandbytes==0.42.0  # 8-bit quantization
#
# ---------------------------------------------------------------------------
# LLM INTEGRATION
# ---------------------------------------------------------------------------
ollama>=0.4.4  # Local LLM (recommended for Windows)

# ---------------------------------------------------------------------------
# PRODUCTION API (FastAPI)
# ---------------------------------------------------------------------------
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pydantic>=2.10.0

# Authentication
python-jose>=3.3.0
passlib[argon2]>=1.7.4
cryptography>=41.0.0
python-multipart>=0.0.6

# ---------------------------------------------------------------------------
# FRONTEND UI
# ---------------------------------------------------------------------------
streamlit>=1.52.0

# ============================================================================
# DEVELOPMENT TOOLS (uncomment if needed)
# ============================================================================
#
# # Interactive development
# jupyter==1.0.0
# ipython==8.20.0
#
# # Testing
# pytest==8.0.0
# pytest-cov==4.1.0
#
# # Code quality
# black==24.1.1
# ruff==0.1.15
#
# ============================================================================
